add lca and calculate edge path length benniehan 12_9
