+++ test/length_metric.json
@@ -1,4 +1,4 @@
 {
-  "method": 2,
+  "method": 3,
   "thereshold": "default"
 }
